requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.1
gunicorn==23.0.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
